fix: warn about loadNodeContent should not be used
This plugin does *not* download file contents. However, some transformer
plugins may call loadNodeContents for S3Asset nodes (because of their
contentType).

This commit adds a warning about it and just returns an empty string.
This way people's builds should not be failing. In the future we can see
how to supress these warnings if they get annoying.